requirements.txt
# Core
PySide6>=6.6
pandas>=2.0
numpy>=1.24
python-dateutil>=2.8
openpyxl>=3.1

# AVRO decoding (prefer fastavro; avro-python3 as fallback)
fastavro>=1.9
avro-python3>=1.10

# Optional but recommended for saving/previewing extra formats
scipy>=1.10      # for MAT output
h5py>=3.10       # for HDF5 preview in the optional viewer
